# internal/exchange/docs/swagger.yaml
basePath: /exchange
definitions:
  github_com_DIMO-Network_token-exchange-api_internal_models.EventFilter:
    properties:
      eventType:
        type: string
      ids:
        items:
          type: string
        type: array
      source:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  internal_controllers_httpcontroller.CloudEvents:
    properties:
      events:
        items:
          $ref: '#/definitions/github_com_DIMO-Network_token-exchange-api_internal_models.EventFilter'
        type: array
    type: object
  internal_controllers_httpcontroller.TokenRequest:
    properties:
      asset:
        description: Asset DID of the asset that permissions are being requested for
          currently either did:erc721 or did:ethr
        type: string
      audience:
        description: Audience is the intended audience for the token.
        items:
          type: string
        type: array
      cloudEvents:
        allOf:
        - $ref: '#/definitions/internal_controllers_httpcontroller.CloudEvents'
        description: CloudEvents contains requests for access to CloudEvents attached
          to the specified NFT.
      nftContractAddress:
        description: |-
          NFTContractAddress is the address of the NFT contract. Privileges will be checked
          on-chain at this address. Address must be in the 0x format e.g. 0x5FbDB2315678afecb367f032d93F642f64180aa3.
          Varying case is okay.
          If asset is provided, this is ignored.
          Deprecated: Use Asset instead.
        example: 0xbA5738a18d83D41847dfFbDC6101d37C69c9B0cF
        type: string
      permissions:
        description: Permissions is a list of the desired permissions.
        items:
          type: string
        type: array
      privileges:
        description: |-
          Privileges is a list of the desired privileges. It must not be empty.
          If Permissions are provided, this is ignored.
          Deprecated: Use Permissions instead.
        example:
        - 1
        - 2
        - 3
        - 4
        items:
          type: integer
        type: array
      tokenId:
        description: |-
          TokenID is the NFT token id.
          If asset is provided, this is ignored.
          Deprecated: Use Asset instead.
        example: 7
        type: integer
    type: object
  internal_controllers_httpcontroller.TokenResponse:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
  title: DIMO Token Exchange API
  version: "1.0"
paths:
  /tokens/exchange:
    post:
      consumes:
      - application/json
      description: Returns a signed token with the requested privileges.
      parameters:
      - description: 'Requested privileges: must include address, token id, and privilege
          ids'
        in: body
        name: tokenRequest
        required: true
        schema:
          $ref: '#/definitions/internal_controllers_httpcontroller.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers_httpcontroller.TokenResponse'
      security:
      - BearerAuth: []
      summary: privileges on the correct token.
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"